Cache the CONNECT_RESUME_ON_READ_WRITE and CONNECT_DATA_IDEMPOTENT values as class constants in Native
